+++ source/codegen/metadata/niscope/CHANGES.md
@@ -64,6 +64,7 @@ The following metadata was added:
 - `'java_package': 'com.ni.grpc.scope'`
 - `'csharp_namespace': 'NationalInstruments.Grpc.Scope'`
 - `'namespace_component': 'niscope'`
+- 'linux_rt_support': True
 
 The fields of the custom waveform-info struct were added to `'custom_types'` in place of the `'ctypes_type'`, `'file_name'`, and `'python_name'`. A  `'grpc_name'` field was also added which was the snake case version of the name field.